+++ cmake/toolchain-arm-clang.cmake
@@ -1,7 +1,7 @@
 #[[
 MIT License
 
-Copyright (c) 2023 ZIMO Elektronik
+Copyright (c) 2024 ZIMO Elektronik
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -27,17 +27,17 @@ SOFTWARE.
 set(CMAKE_TRY_COMPILE_TARGET_TYPE STATIC_LIBRARY)
 set(CMAKE_SYSTEM_NAME Target)
 
-# Find arm-none-eabi
-find_program(ARM_NONE_EABI_C_COMPILER arm-none-eabi-gcc)
-find_program(ARM_NONE_EABI_CXX_COMPILER arm-none-eabi-g++)
+# Find arm-none-eabi-gcc
+find_program(ARM_NONE_EABI_C_COMPILER arm-none-eabi-gcc REQUIRED)
+find_program(ARM_NONE_EABI_CXX_COMPILER arm-none-eabi-g++ REQUIRED)
 
 # Find clang
-find_program(C_COMPILER clang)
-find_program(CXX_COMPILER clang++)
-find_program(AR llvm-ar)
-find_program(OBJCOPY llvm-objcopy)
-find_program(OBJDUMP llvm-objdump)
-find_program(SIZE llvm-size)
+find_program(C_COMPILER clang REQUIRED)
+find_program(CXX_COMPILER clang++ REQUIRED)
+find_program(AR llvm-ar REQUIRED)
+find_program(OBJCOPY llvm-objcopy REQUIRED)
+find_program(OBJDUMP llvm-objdump REQUIRED)
+find_program(SIZE llvm-size REQUIRED)
 
 set(CMAKE_ASM_COMPILER ${C_COMPILER})
 set(CMAKE_C_COMPILER ${C_COMPILER})