+++ cmake/version_from_git.cmake
@@ -1,7 +1,7 @@
 #[[
 MIT License
 
-Copyright (c) 2023 ZIMO Elektronik
+Copyright (c) 2024 ZIMO Elektronik
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -23,7 +23,7 @@ SOFTWARE.
 ]]
 
 function(version_from_git)
-  find_package(Git)
+  find_package(Git REQUIRED)
 
   # Run git describe --tags
   execute_process(