+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+##
+- Add get_qt
+- Add VERSION argument to get_cqtdeployer
+- Remove x86_64-w64-mingw32 toolchain file
+
 ## 0.7.0
 - Add add_compile_link_options
 - CPM gets added implicitly